wavvol wave limits

+++ wavvol.c
@@ -8,7 +8,7 @@ int main(int argc, char **argv){
 
 	/*checks the possibles flags*/
 	char *input_flag=NULL, *output_flag=NULL;
-	float level=1;
+	double level=1;
 	flags(&input_flag, &output_flag, &level, argc, argv);
 
 	FILE *input = checksInput(input_flag);
